+++ src/patient/entities/patient.entity.ts
@@ -33,16 +33,16 @@ export class Patient {
   @Column()
   admitted: string;
 
-  @Column('int', { default: 0 })
+  @Column('float', { default: 0 })
   bedHeadPosition: number;
 
-  @Column('int', { default: 0 })
+  @Column('float', { default: 0 })
   bedLeftPosition: number;
 
-  @Column('int', { default: 0 })
+  @Column('float', { default: 0 })
   bedRightPosition: number;
 
-  @Column('int', { default: 0 })
+  @Column('float', { default: 0 })
   bedTiltPosition: number;
 
   @ManyToOne(() => User, (user) => user.patients, {
